+++ src/main/java/ru/practicum/shareit/user/User.java
@@ -3,6 +3,8 @@ package ru.practicum.shareit.user;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 @Builder
@@ -14,7 +16,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_id")
-    private Integer id;
+    private BigInteger id;
     private String name;
     private String email;
 }
