+++ src/main/java/ru/practicum/shareit/user/UserController.java
@@ -6,6 +6,8 @@ import org.springframework.web.bind.annotation.*;
 import ru.practicum.shareit.user.dto.UpdateUserDto;
 import ru.practicum.shareit.user.dto.UserDto;
 
+import java.math.BigInteger;
+
 @RestController
 @RequestMapping(path = "/users")
 @RequiredArgsConstructor
@@ -13,7 +15,7 @@ public class UserController {
     private final UserService userService;
 
     @GetMapping("/{userId}")
-    public UserDto getUserById(@PathVariable Integer userId) {
+    public UserDto getUserById(@PathVariable BigInteger userId) {
         return userService.getUserById(userId);
     }
 
@@ -23,12 +25,12 @@ public class UserController {
     }
 
     @PatchMapping("/{userId}")
-    public UserDto updateUser(@Valid @RequestBody UpdateUserDto userDto, @PathVariable Integer userId) {
+    public UserDto updateUser(@Valid @RequestBody UpdateUserDto userDto, @PathVariable BigInteger userId) {
         return userService.updateUser(userDto, userId);
     }
 
     @DeleteMapping("/{userId}")
-    public void deleteUserById(@PathVariable Integer userId) {
+    public void deleteUserById(@PathVariable BigInteger userId) {
         userService.deleteUserById(userId);
     }
 }
