+++ src/main/java/ru/practicum/shareit/user/UserService.java
@@ -3,12 +3,14 @@ package ru.practicum.shareit.user;
 import ru.practicum.shareit.user.dto.UpdateUserDto;
 import ru.practicum.shareit.user.dto.UserDto;
 
+import java.math.BigInteger;
+
 public interface UserService {
-    UserDto getUserById(Integer userId);
+    UserDto getUserById(BigInteger userId);
 
     UserDto createUser(UserDto userDto);
 
-    UserDto updateUser(UpdateUserDto userDto, Integer userId);
+    UserDto updateUser(UpdateUserDto userDto, BigInteger userId);
 
-    void deleteUserById(Integer userId);
+    void deleteUserById(BigInteger userId);
 }
