+++ src/main/java/ru/practicum/shareit/user/UserStorage.java
@@ -2,8 +2,8 @@ package ru.practicum.shareit.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
+import java.math.BigInteger;
 
-public interface UserStorage extends JpaRepository<User, Integer> {
+public interface UserStorage extends JpaRepository<User, BigInteger> {
     boolean existsByEmail(String email);
 }
