+++ src/main/java/ru/practicum/shareit/user/dto/UserDto.java
@@ -8,12 +8,14 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 @Builder
 public class UserDto {
     @JsonProperty(access = JsonProperty.Access.READ_ONLY)
-    private Integer id;
+    private BigInteger id;
     @NotBlank(message = "Имя не может быть пустым")
     private String name;
     @NotNull(message = "email должен быть задан")
